Moved converter functions into extensions. Changed the MachineMutator extension into the converter struct definition.
